requirements.txt
beautifulsoup4==4.12.3
dash==2.17.1
dash_bootstrap_components==1.6.0
numpy==2.0.0
pandas==2.2.2
Pillow==10.3.0
plotly==5.22.0
Requests==2.32.3
tqdm==4.65.0
urllib3==2.2.2